pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eco-draft-2d"
version = "0.1.0"
description = "FastAPI backend for co-creative 2D CAD + engineering diagrams with sustainability metrics"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Eco Draft Team", email = "team@ecodraft.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "shapely>=2.0.0",
    "ezdxf>=1.1.0",
    "svgwrite>=1.4.0",
    "reportlab>=4.0.0",
    "pymoo>=0.6.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "requests>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.6.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://github.com/eco-draft/eco-draft-2d"
Documentation = "https://github.com/eco-draft/eco-draft-2d#readme"
Repository = "https://github.com/eco-draft/eco-draft-2d.git"
Issues = "https://github.com/eco-draft/eco-draft-2d/issues"

[tool.setuptools.packages.find]
where = ["backend"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "ezdxf.*",
    "svgwrite.*",
    "reportlab.*",
    "pymoo.*",
]
ignore_missing_imports = true
